+++ src/lib.rs
@@ -19,7 +19,7 @@ mod brush;
 mod cache;
 mod error;
 mod pipeline;
-
+// TODO remove anything about scissor regions and set_load_op
 pub use brush::{BrushBuilder, TextBrush};
 
 /// Contains all needed objects for inserting, styling and iterating text and glyphs.